chore: 1-click refactor for workspace service
