+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/WorkspaceService.java
@@ -1,5 +1,5 @@
 package com.appsmith.server.services;
 
-import com.appsmith.server.services.ce.WorkspaceServiceCE;
+import com.appsmith.server.services.ce_compatible.WorkspaceServiceCECompatible;
 
-public interface WorkspaceService extends WorkspaceServiceCE {}
+public interface WorkspaceService extends WorkspaceServiceCECompatible {}
