+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce_compatible/WorkspaceServiceCECompatible.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.services.ce_compatible;
+
+import com.appsmith.server.services.ce.WorkspaceServiceCE;
+
+public interface WorkspaceServiceCECompatible extends WorkspaceServiceCE {}
